backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
pyjwt==2.8.0
twilio==9.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-cloud-storage==2.14.0
Pillow==10.2.0
marshmallow==3.21.0
marshmallow-sqlalchemy==0.29.0

